requirements.txt
google-generativeai>=0.7.0
huggingface_hub>=0.24.6
moviepy==1.0.3
streamlit==1.39.0
python-dotenv==1.0.1
Pillow==10.4.0



